+++ tests/test_resolver.py
@@ -101,7 +101,10 @@ import pytest
         (['Flask', ('click', True), ('itsdangerous', True)],
          ['flask==0.10.1', 'itsdangerous==0.24', 'markupsafe==0.23',
           'jinja2==2.7.3', 'werkzeug==0.10.4']
-         )
+         ),
+
+        # Exclude package dependcy of setuptools as it is unsafe.
+        (['html5lib'], ['html5lib==0.999999999']),
     ])
 )
 def test_resolver(resolver, from_line, input, expected, prereleases):