+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fix duplicate entries that could happen in generated requirements.txt (#427)
 - Gracefully report invalid pip version (#457)
 - Fix capitalization in the generated requirements.txt, packages will always be lowercased (#452)
+- Exclude unsafe packages' dependencies when `--allow-unsafe` is not in use (#445)
 
 
 # 1.8.0